Gets accounts from database by date rough draft
The query that gets the accountsInfo from the database gets a querysnapshot everytime, no matter the provided info. This is an issue because it should be providing an error when the info is not found. I made a temporary fix of counting the length of the querysnapshot query and checking if it's more than 0 before I continue to filling in the info on the front end. Works but I need to refine the query